+++ packages/instrument-utils/src/guards/isSepaInstrument/isSepaInstrument.ts
@@ -1,0 +1,8 @@
+import { BankInstrument, PaymentInstrument } from '@bigcommerce/checkout-sdk';
+
+export function isSepaInstrument(instrument?: PaymentInstrument): instrument is BankInstrument {
+    return (
+        instrument?.type === 'bank' &&
+        (instrument?.method === 'sepa' || instrument?.method === 'sepa_direct_debit')
+    );
+}