+++ packages/paypal-connect-integration/src/is-paypal-commerce-connect-method.test.ts
@@ -4,7 +4,7 @@ import isPayPalCommerceConnectMethod from './is-paypal-commerce-connect-method';
 
 describe('isPaypalCommerceConnectMethod', () => {
     it('returns true if provided methodId is related to PayPal Connect', () => {
-        expect(isPayPalCommerceConnectMethod(PaymentMethodId.PaypalCommerce)).toBe(true);
+        expect(isPayPalCommerceConnectMethod(PaymentMethodId.PaypalCommerceCreditCards)).toBe(true);
         expect(
             isPayPalCommerceConnectMethod(PaymentMethodId.PayPalCommerceAcceleratedCheckout),
         ).toBe(true);